Addressing panic from issue #223
The approach here was to utilize UnmarshalJSON() on the Command domain
model in order to ensure that both Get and Put are initialized. The
bug was caused because the sample request did not have a GET so when
Mongo tried to call Get.String() it panicked due to NIL reference.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>